Decomposite component structure down to BookShelf and Book
Here also introduced the book shelf update. The source of truth is 
BooksApp state. The change shelf request is sent upward from Book to 
BooksApp, the books repository is updated there with BooksAPI.update(), 
the state is changed with the books refreshed by the BooksAPI.getAll(), 
and the change is replicated across the entire app by the React.